refactor(templates): use named route boundaries
