+++ templates/react-router/app/routes/products_.layout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link, Outlet, useLocation } from "react-router";
-import { SsgoiTransitionBoundary } from "../components/ssgoi-transition-boundary";
+import { SsgoiRouteBoundary } from "../components/ssgoi-route-boundary";
 import { PRODUCT_CATEGORIES } from "../components/ssgoi-config";
 
 export default function ProductsLayout() {
@@ -8,8 +8,8 @@ export default function ProductsLayout() {
   const pathname = location.pathname;
 
   return (
-    <SsgoiTransitionBoundary
-      scope={() => "products-layout"}
+    <SsgoiRouteBoundary
+      name="products-shell"
       className="min-h-screen bg-[#121212] flex flex-col"
     >
       {/* Header - Fixed */}
@@ -37,10 +37,10 @@ export default function ProductsLayout() {
 
       {/* Tab Content - Slide transitions here */}
       <div className="flex-1 overflow-hidden relative">
-        <SsgoiTransitionBoundary className="min-h-full bg-[#121212]">
+        <SsgoiRouteBoundary name="page" className="min-h-full bg-[#121212]">
           <Outlet />
-        </SsgoiTransitionBoundary>
+        </SsgoiRouteBoundary>
       </div>
-    </SsgoiTransitionBoundary>
+    </SsgoiRouteBoundary>
   );
 }
