+++ templates/tanstack-router/README.md
@@ -8,25 +8,28 @@ pnpm dev
 ## Structure
 
 - `app/components/ssgoi-config.ts`: one transition config.
-- `app/components/demo-layout.tsx`: one root `<Ssgoi>`.
-- `app/components/ssgoi-transition-boundary.tsx`: reads router state.
-- Parent route files place boundaries around `<Outlet />`.
+- `app/components/demo-layout.tsx`: one root `<Ssgoi>` inside a
+  `relative z-0 overflow-x-clip` shell.
+- `app/components/ssgoi-route-boundary.tsx`: name → route id/key resolver
+  using router state.
+- Parent route files select boundary names around `<Outlet />`.
 
 ```tsx
 const pathname = useRouterState({
   select: (state) => state.location.pathname,
 });
+const boundary = resolveBoundary(name, pathname);
 
 return (
-  <div key={scope(pathname)} data-ssgoi-transition={pathname}>
+  <div key={boundary.key} data-ssgoi-transition={boundary.id}>
     {children}
   </div>
 );
 ```
 
-The products parent route keeps its outer layout key stable and keys only the
-inner content by pathname. Ordered paths in the root config decide slide
-direction.
+`page` uses the pathname for both values. `products-shell` keeps the
+route-owned outer layout key stable and the inner content uses `page`. Ordered
+paths in the root config decide slide direction.
 
 Effects:
