+++ templates/tanstack-router/app/routes/profile.tsx
@@ -1,10 +1,10 @@
 import { createFileRoute, Outlet } from "@tanstack/react-router";
-import { SsgoiTransitionBoundary } from "../components/ssgoi-transition-boundary";
+import { SsgoiRouteBoundary } from "../components/ssgoi-route-boundary";
 
 export const Route = createFileRoute("/profile")({
   component: () => (
-    <SsgoiTransitionBoundary className="min-h-full bg-[#121212]">
+    <SsgoiRouteBoundary name="page" className="min-h-full bg-[#121212]">
       <Outlet />
-    </SsgoiTransitionBoundary>
+    </SsgoiRouteBoundary>
   ),
 });
